# CM.Portal/src/app/api/models/job-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface JobDto {
  cowId?: number | null;
  date?: string;
  description?: string | null;
  id?: number;
  infos?: string | null;
  penId?: number | null;
  status?: string | null;
  title?: string | null;
}